+++ public/robots.txt
@@ -1,2 +1,4 @@
 User-Agent: *
-Disallow:
+Disallow: /editor
+Disallow: /editor/
+Disallow: /editor/*
